Move BeansAPIRelationshipIterator out of the compiler
This will remove some kernel dependencies from the cypher compiler.
